move plot_orientation to bearing module
